requirements.txt
Django==5.0.3
djangorestframework==3.15.1
psycopg2-binary==2.9.9
celery==5.3.6
redis==5.0.3
django-redis==5.4.0
gunicorn==21.2.0
django-ratelimit==4.1.0
whitenoise==6.6.0
django-cors-headers==4.3.1
python-dotenv==1.0.0